backend/pyproject.toml
[tool.poetry]
name = "energisa-voice-backend"
version = "1.0.0"
description = "Backend Python do agente de voz GISA"
authors = ["Energisa"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-dotenv = "^1.0.0"
livekit = "^0.11.0"
livekit-api = "^0.6.0"
deepgram-sdk = "^3.4.0"
google-generativeai = "^0.3.2"
elevenlabs = "^0.2.27"
aiohttp = "^3.9.1"
websockets = "^12.0"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"

[tool.poetry.dev-dependencies]
black = "^23.12.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
